playoff braket update

+++ pkg/data/sleeper.go
@@ -156,7 +156,7 @@ type SleeperLeague struct {
 	PreviousLeagueID      string                       `json:"previous_league_id"`
 	Name                  string                       `json:"name"`
 	Metadata              SleeperLeagueMetadata        `json:"metadata"`
-	LoserBracketID        int                          `json:"loser_bracket_id"`
+	LoserBracketID        int64                          `json:"loser_bracket_id"`
 	LeagueID              string                       `json:"league_id"`
 	LastReadID            string                       `json:"last_read_id"`
 	LastPinnedMessageID   string                       `json:"last_pinned_message_id"`
@@ -171,7 +171,7 @@ type SleeperLeague struct {
 	GroupID               string                       `json:"group_id"`
 	DraftID               string                       `json:"draft_id"`
 	CompanyID             string                       `json:"company_id"`
-	BracketID             int                          `json:"bracket_id"`
+	BracketID             int64                          `json:"bracket_id"`
 	Avatar                interface{}                  `json:"avatar"`
 }
